+++ baybe/recommenders/pure/bayesian/botorch.py
@@ -1,13 +1,19 @@
 """Botorch recommender."""
 
+from __future__ import annotations
+
 import math
-from typing import Any, ClassVar
+from collections.abc import Collection
+from typing import TYPE_CHECKING, Any, ClassVar
 
 import pandas as pd
 from attr.converters import optional
 from attrs import define, field
+from attrs.validators import ge, instance_of
 
+from baybe.constraints import ContinuousCardinalityConstraint
 from baybe.exceptions import NoMCAcquisitionFunctionError
+from baybe.parameters.numerical import _FixedNumericalContinuousParameter
 from baybe.recommenders.pure.bayesian.base import BayesianRecommender
 from baybe.searchspace import (
     SearchSpace,
@@ -20,6 +26,9 @@ from baybe.utils.sampling_algorithms import (
     DiscreteSamplingMethod,
     sample_numerical_df,
 )
+
+if TYPE_CHECKING:
+    from torch import Tensor
 
 
 @define(kw_only=True)
@@ -57,6 +66,16 @@ class BotorchRecommender(BayesianRecommender):
     sampling_percentage: float = field(default=1.0)
     """Percentage of discrete search space that is sampled when performing hybrid search
     space optimization. Ignored when ``hybrid_sampler="None"``."""
+
+    n_threshold_inactive_parameters_generator: int = field(
+        default=10, validator=[instance_of(int), ge(1)]
+    )
+    """Threshold used for checking which inactive parameters generator is used when
+    cardinality constraints are present. When the size of the combinatorial list of
+    all possible inactive parameters is larger than the threshold, a fixed number of
+    randomly generated inactive parameter configurations are used and the best
+    optimum among them is recommended; Otherwise, we find the best one by iterating the
+    combinatorial list of all possible inactive parameters """
 
     @sampling_percentage.validator
     def _validate_percentage(  # noqa: DOC101, DOC103
@@ -151,31 +170,174 @@ class BotorchRecommender(BayesianRecommender):
                 f"acquisition functions for batch sizes > 1."
             )
 
+        if len(subspace_continuous.constraints_cardinality):
+            points, _ = self._recommend_continuous_with_cardinality_constraints(
+                subspace_continuous,
+                batch_size,
+            )
+        else:
+            points, _ = self._recommend_continuous_without_cardinality_constraints(
+                subspace_continuous,
+                batch_size,
+            )
+
+        # Return optimized points as dataframe
+        rec = pd.DataFrame(points, columns=subspace_continuous.param_names)
+        return rec
+
+    def _recommend_continuous_with_cardinality_constraints(
+        self,
+        subspace_continuous: SubspaceContinuous,
+        batch_size: int,
+    ) -> tuple[Tensor, Tensor]:
+        """Recommend from a continuous search space with cardinality constraints.
+
+        Args:
+            subspace_continuous: The continuous subspace from which to generate
+                recommendations.
+            batch_size: The size of the recommendation batch.
+
+        Returns:
+            The recommendations.
+            The acquisition values.
+
+        Raises:
+            RuntimeError: If the continuous search space has no cardinality constraint.
+        """
+        import torch
+
+        if not subspace_continuous.constraints_cardinality:
+            raise RuntimeError(
+                f"This method expects a subspace object with constraints of type "
+                f"{ContinuousCardinalityConstraint.__name__}. For a subspace object "
+                f"without constraints of type"
+                f" {ContinuousCardinalityConstraint.__name__}, use method"
+                f"{self._recommend_continuous_without_cardinality_constraints.__name__}."  # noqa
+            )
+
+        acqf_values_all: list[Tensor] = []
+        points_all: list[Tensor] = []
+
+        def append_recommendation_for_inactive_parameters_setting(
+            inactive_parameters: Collection[str],
+        ):
+            """Append the recommendation for each inactive parameter configuration.
+
+            Args:
+                inactive_parameters: A list of inactive parameters.
+            """
+            # Create a new subspace by ensuring all active parameters being
+            # non-zeros.
+            subspace_continuous_without_cardinality_constraints = (
+                subspace_continuous._remove_cardinality_constraints(inactive_parameters)
+            )
+            try:
+                # Optimize the acquisition function
+                (
+                    points_i,
+                    acqf_values_i,
+                ) = self._recommend_continuous_without_cardinality_constraints(
+                    subspace_continuous_without_cardinality_constraints,
+                    batch_size,
+                )
+                # Append recommendation list and acquisition function values
+                points_all.append(points_i.unsqueeze(0))
+                acqf_values_all.append(acqf_values_i.unsqueeze(0))
+
+            # The optimization problem may be infeasible for certain inactive
+            # parameters. The optimize_acqf raises a ValueError when the optimization
+            # problem is infeasible.
+            except ValueError:
+                pass
+
+        # Below we start recommendation
+        if (
+            subspace_continuous.n_inactive_parameter_combinations
+            > self.n_threshold_inactive_parameters_generator
+        ):
+            # When the combinatorial list is too large, randomly set some parameters
+            # inactive.
+            for _ in range(self.n_threshold_inactive_parameters_generator):
+                inactive_params_sample = tuple(
+                    subspace_continuous._sample_inactive_parameters(1)[0]
+                )
+                append_recommendation_for_inactive_parameters_setting(
+                    inactive_params_sample
+                )
+        else:
+            # When the combinatorial list is not too large, iterate the combinatorial
+            # list of all possible inactive parameters.
+            for (
+                inactive_params_generator
+            ) in subspace_continuous.inactive_parameter_combinations():
+                append_recommendation_for_inactive_parameters_setting(
+                    inactive_params_generator
+                )
+
+        # Find the best option
+        points = torch.cat(points_all)[torch.argmax(torch.cat(acqf_values_all)), :]
+        acqf_values = torch.max(torch.cat(acqf_values_all))
+        return points, acqf_values
+
+    def _recommend_continuous_without_cardinality_constraints(
+        self,
+        subspace_continuous: SubspaceContinuous,
+        batch_size: int,
+    ) -> tuple[Tensor, Tensor]:
+        """Recommend from a continuous search space without cardinality constraints.
+
+        Args:
+            subspace_continuous: The continuous subspace from which to generate
+                recommendations.
+            batch_size: The size of the recommendation batch.
+
+        Returns:
+            The recommendations.
+            The acquisition values.
+
+        Raises:
+            RuntimeError: If the continuous search space has any cardinality
+                constraints.
+        """
         import torch
         from botorch.optim import optimize_acqf
 
-        points, _ = optimize_acqf(
+        if subspace_continuous.constraints_cardinality:
+            raise RuntimeError(
+                f"This method expects only subspace object without constraints of type "
+                f"{ContinuousCardinalityConstraint.__name__}. For a subspace object "
+                f"with constraints of type {ContinuousCardinalityConstraint.__name__}, "
+                f"try method {self._recommend_continuous.__name__}."
+            )
+
+        fixed_parameters = {
+            idx: p.value
+            for (idx, p) in enumerate(subspace_continuous.parameters)
+            if isinstance(p, _FixedNumericalContinuousParameter)
+        }
+
+        points, acqf_values = optimize_acqf(
             acq_function=self._botorch_acqf,
             bounds=torch.from_numpy(subspace_continuous.param_bounds_comp),
             q=batch_size,
             num_restarts=5,  # TODO make choice for num_restarts
             raw_samples=10,  # TODO make choice for raw_samples
+            fixed_features=fixed_parameters or None,
             equality_constraints=[
                 c.to_botorch(subspace_continuous.parameters)
                 for c in subspace_continuous.constraints_lin_eq
             ]
-            or None,  # TODO: https://github.com/pytorch/botorch/issues/2042
+            or None,
+            # TODO: https://github.com/pytorch/botorch/issues/2042
             inequality_constraints=[
                 c.to_botorch(subspace_continuous.parameters)
                 for c in subspace_continuous.constraints_lin_ineq
             ]
-            or None,  # TODO: https://github.com/pytorch/botorch/issues/2042
+            or None,
+            # TODO: https://github.com/pytorch/botorch/issues/2042
             sequential=self.sequential_continuous,
         )
-
-        # Return optimized points as dataframe
-        rec = pd.DataFrame(points, columns=subspace_continuous.param_names)
-        return rec
+        return points, acqf_values
 
     def _recommend_hybrid(
         self,